Changed dictionary structure so that labels and data are at the same level in the channel
